# docs/openspec/registry.openapi.yaml
openapi: 3.0.3
info:
  title: Virbius Control Plane API (MVP)
  version: 1.3.0-mvp
  description: |
    **virbius-control** 单 Java 进程（Registry + Admin 合并）。

    - **规范面**：`/api/v1/admin/tenants/{tenantId}/...`，响应 `{ code, message, data, timestamp }`（`code=0` 成功）。
    - **兼容面**：`/api/v1/tenants/{tenantId}/...`（Legacy，裸 JSON），逐步废弃。
    - 规则真源：`rule_history`；`runtime` 见 DESIGN §8.5.0。
    - **规则级 enforce / canary**（管侧真拦 + 云 PolicyMerge）：见 [rule-level-enforce.md](./rule-level-enforce.md)。
    - PoC 发布为 **同步 HTTP 200**（非 202 异步）。
    - Bundle YAML 导入（`draft/import`）**未实现**，Post-MVP。
servers:
  - url: http://virbius-control:8080
    description: PoC（Registry + Admin UI 同端口）
...
paths:
...
  /api/v1/admin/tenants/{tenantId}/rules:
    get:
      tags: [admin-rules]
      summary: 列出租户规则（当前 revision 指针）
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - name: layer
          in: query
          schema:
            type: string
            enum: [edge, gateway, cloud]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultRulePointerList'
    post:
      tags: [admin-rules]
      summary: 创建或更新规则（追加 rule_history revision）
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertRuleRequest'
      responses:
        '200':
          description: 新 revision 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultRuleRevision'
  /api/v1/admin/tenants/{tenantId}/rules/{ruleId}:
    get:
      tags: [admin-rules]
      summary: 获取规则当前 revision
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultRuleRevision'
  /api/v1/admin/tenants/{tenantId}/rules/{ruleId}/revisions:
    get:
      tags: [admin-rules]
      summary: 列出规则历史 revision
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultRuleRevisionMetaList'
  /api/v1/admin/tenants/{tenantId}/rules/{ruleId}/revisions/{revision}:
    get:
      tags: [admin-rules]
      summary: 获取指定 revision 全文
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
        - $ref: '#/components/parameters/revision'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultRuleRevision'
  /api/v1/admin/tenants/{tenantId}/rules/{ruleId}/status:
    patch:
      tags: [admin-rules]
      summary: 启用/停用规则（逻辑删除，追加 revision）
      description: |
        `disabled` 的规则不进 gateway/edge 产物与 RuleCache；**不可**再改 body/enforce。
        变更后自动 sync 名单产物并刷新 engine。
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuleStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultRuleRevision'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/tenants/{tenantId}/rules/{ruleId}/runtime:
    patch:
      tags: [admin-rules]
      summary: 规则级放量（更新 enforce_mode，追加 rule_history revision）
      description: |
        **规则级** enforce / canary（非租户 / Bundle 级）。详见 [rule-level-enforce.md](./rule-level-enforce.md)。

        | enforce_mode | canary_percent | 行为（在规则所属 layer 执行点） |
        |--------------|----------------|----------------------------------|
        | `dry_run` | 忽略 | 命中 `effective_action=review`，不 403/428 |
        | `canary` | **必填** 1–100 | 命中且 session 在桶内：`block`/`captcha`；桶外：`review` |
        | `full` | 忽略 | 命中即 `block`/`captcha`（由 `intent_action` 决定） |

        **合并**：同层按 `intent_action` 优先级（deny > captcha > review > allow），再应用 `enforce_mode`。详见 [rule-level-enforce.md](./rule-level-enforce.md)。

        **副作用**：
        - `layer=gateway`：刷新 gateway 名单产物（`refreshArtifacts` / `sync-rules`），enforce 写入 `{tenant}-access-lists.json`。
        - `layer=cloud`：刷新 engine RuleCache（`runtime/publish-snapshot` 或 publish 流程）。
        - gateway 规则 **终局 403** 时不调用 evaluate（管侧审计必写）。

        新建规则默认 `dry_run`；**不应**在 `POST .../rules` 时设置 canary，仅在本接口放量。
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuntimeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultRuleRevision'
  /api/v1/admin/tenants/{tenantId}/rules/{ruleId}/runtime/publish-snapshot:
    post:
      tags: [admin-publish]
      summary: 刷新租户 engine RuleCache（runtime_only）
      description: |
        **PoC：`ruleId` 路径参数被忽略**，始终刷新整租户 RuleCache（不编译 gateway/edge）。

        适用于 **cloud 层** 规则 `enforce_mode` / `canary_percent` 变更。
        **gateway 层** enforce 变更须刷新 gateway 产物（`POST .../access-lists/sync-rules` 或保存 gateway 规则触发的 artifact sync），见 [rule-level-enforce.md](./rule-level-enforce.md)。
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      responses:
        '200':
          description: EngineCacheSync 完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'

  # --- Legacy（兼容脚本；新集成请用 Admin）---
...
  /api/v1/tenants/{tenantId}/rules/{ruleId}:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 获取规则"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleRevision'
  /api/v1/tenants/{tenantId}/rules/{ruleId}/runtime:
    patch:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 规则级放量（enforce_mode / canary_percent）"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuntimeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleRevision'
components:
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        example: default
...
    ruleId:
      name: ruleId
      in: path
      required: true
      schema:
        type: string
    revision:
      name: revision
      in: path
      required: true
      schema:
        type: integer
        format: int32
...
  responses:
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResultError'
  schemas:
...
    ApiResultError:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
    ApiResultObject:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
        data:
          type: object
          additionalProperties: true
...
    ApiResultRulePointerList:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/RulePointer'
    ApiResultRuleRevision:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/RuleRevision'
    ApiResultRuleRevisionMetaList:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/RuleRevisionMeta'
...
    RulePointer:
      type: object
      properties:
        rule_id:
          type: string
        current_revision:
          type: integer
        layer:
          type: string
        runtime:
          type: string
          enum: [lua-dsl, lua, prompt, groovy, list_match, cumulative]
        reason_code:
          type: string
        intent_action:
          type: string
          enum: [allow, deny, captcha, review]
          default: deny
          description: 规则命中后的希望动作；见 rule-level-enforce.md
        enforce_mode:
          type: string
          enum: [dry_run, canary, full]
          description: 规则级执行模式；见 rule-level-enforce.md
        canary_percent:
          type: integer
          minimum: 1
          maximum: 100
          nullable: true
          description: enforce_mode=canary 时必填；full/dry_run 时应省略或 null
        rule_status:
          type: string
          enum: [draft, active, disabled]
          description: |
            draft=配置中不进执行面；active=参与匹配；disabled=停用不可编辑。
            流转：draft→active|disabled；active→disabled；disabled→draft。
    UpdateRuleStatusRequest:
      type: object
      required: [rule_status]
      properties:
        rule_status:
          type: string
          enum: [draft, active, disabled]
          description: |
            draft=配置中不进执行面；active=参与匹配；disabled=停用不可编辑。
            流转：draft→active|disabled；active→disabled；disabled→draft。
    UpsertRuleRequest:
      type: object
      required: [rule_id, layer, runtime, reason_code, body]
      properties:
        rule_id:
          type: string
        bundle_id:
          type: string
        layer:
          type: string
          enum: [edge, gateway, cloud]
        runtime:
          type: string
          enum: [lua-dsl, lua, prompt, groovy, list_match, cumulative]
          description: 详见 DESIGN.md §8.5.0
        reason_code:
          type: string
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
          default: 100
        intent_action:
          type: string
          enum: [allow, deny, captcha, review]
          default: deny
          description: 规则命中后的希望动作
        scope:
          type: object
        body:
          oneOf:
            - type: string
            - type: object
      description: |
        新建或更新规则正文。**不包含** enforce / canary（默认 dry_run）。
        放量使用 `PATCH .../rules/{ruleId}/runtime`。
    UpdateRuntimeRequest:
      type: object
      required: [enforce_mode]
      properties:
        enforce_mode:
          type: string
          enum: [dry_run, canary, full]
          description: 规则级 enforce；见 rule-level-enforce.md
        canary_percent:
          type: integer
          minimum: 1
          maximum: 100
          nullable: true
          description: enforce_mode=canary 时必填；否则省略或 null
    RuleRevisionMeta:
      type: object
      properties:
        rule_revision:
          type: integer
        enforce_mode:
          type: string
          enum: [dry_run, canary, full]
        canary_percent:
          type: integer
          minimum: 1
          maximum: 100
          nullable: true
        rule_status:
          type: string
          enum: [draft, active, disabled]
          description: |
            draft=配置中不进执行面；active=参与匹配；disabled=停用不可编辑。
            流转：draft→active|disabled；active→disabled；disabled→draft。
        modified_at:
          type: string
          format: date-time
        effective_from:
          type: string
          format: date-time
        effective_to:
          type: string
          format: date-time
          nullable: true
    RuleRevision:
      allOf:
        - $ref: '#/components/schemas/RuleRevisionMeta'
        - type: object
          properties:
            tenant_id:
              type: string
            rule_id:
              type: string
            layer:
              type: string
            runtime:
              type: string
            reason_code:
              type: string
            risk_score:
              type: integer
            intent_action:
              type: string
              enum: [allow, deny, captcha, review]
            scope:
              type: object
            body:
              oneOf:
                - type: string
                - type: object
            body_hash:
              type: string
...